requirements.txt
cvxpy
scipy
numpy
jax
plotly
pyaml
termcolor